Extract CLI argv helpers
